Add solution for Task_2

+++ Semester_1/Homework_5/Task_2/stack.c
@@ -3,30 +3,30 @@
 
 struct StackElement
 {
-    int value;
-    StackElement *next;
+    double value;
+    StackElement* next;
 };
 
 struct Stack
 {
-    StackElement *first;
+    StackElement* first;
 };
 
-Stack *createStack()
+Stack* createStack()
 {
-    Stack *stack = malloc(sizeof(struct Stack));
+    Stack* stack = malloc(sizeof(struct Stack));
     stack->first = NULL;
     return stack;
 }
 
-bool isEmpty(Stack *stack)
+bool isEmpty(Stack* stack)
 {
     return stack->first == NULL;
 }
 
-bool push(int value, Stack *stack)
+bool push(double value, Stack* stack)
 {
-    StackElement *stackElement = malloc(sizeof(struct StackElement));
+    StackElement* stackElement = malloc(sizeof(struct StackElement));
     stackElement->value = value;
     stackElement->next = stack->first;
 
@@ -34,16 +34,16 @@ bool push(int value, Stack *stack)
     return true;
 }
 
-int pop(Stack *stack)
+double pop(Stack* stack)
 {
     if (isEmpty(stack))
     {
         return 0;
     }
 
-    StackElement *poppedElement = stack->first;
+    StackElement* poppedElement = stack->first;
     stack->first = poppedElement->next;
-    int value = poppedElement->value;
+    double value = poppedElement->value;
     free(poppedElement);
     return value;
 }
