+++ Semester_1/Homework_5/Task_2/stack.h
@@ -9,9 +9,9 @@ struct StackElement;
 typedef struct Stack Stack;
 typedef struct StackElement StackElement;
 
-bool isEmpty(Stack *stack);
-bool push(int value, Stack *stack);
-int pop(Stack *stack);
-Stack *createStack();
+bool isEmpty(Stack* stack);
+bool push(double value, Stack* stack);
+double pop(Stack* stack);
+Stack* createStack();
 
 #endif
